Adding code to compute temporal feature matrix and showdown probabilities
A) Temporal feature matrix (TFM) is a [features x time_steps] matrix
that keeps track of the progress of a hand. Can be used as an input to
sequence based nets (e.g. LSTMs). The rows of the TFM are as follows:
1) hero bet amount (cumulative)
2) villain bet amount (cumulative)
3) street (0 = pre flop, 1 flop, 2 turn, 3 river)
4) hero discard? (1 = yes, 0 = no)
5) villain discard? ‘’

Columns are added to this matrix for every action taken in the hand
(e.g. after every blind posting, bet, raise, check, deal, etc).

B) Also added method to compute showdown probabilities based on Amaro’s
work with the Deuces package.